Moved code for storing fields history to nchistory.cpp module
