+++ src/nceditline.cpp
@@ -5,94 +5,15 @@
 */
 
 #include "nceditline.h"
-#include "wcm-config.h"
 #include "string-util.h"
 #include "unicode_lc.h"
+#include "nchistory.h"
 
-
-#define MAX_FIELD_HISTORY_COUNT	50
-
-typedef ccollect<std::vector<unicode_t>> HistCollect;
-
-static cstrhash<clPtr<HistCollect>> g_histHash;
-
-
-inline bool HistoryCanSave()
-{
-	//return wcmConfig.systemSaveHistory;
-	return true;
-}
 
 inline bool AcEnabled()
 {
 	//return wcmConfig.systemAutoComplete;
 	return true;
-}
-
-HistCollect* HistGetList( const char* fieldName )
-{
-	if ( !fieldName || !fieldName[0] )
-	{
-		return nullptr;
-	}
-	
-	clPtr<HistCollect>* pp = g_histHash.exist( fieldName );
-	if ( pp && pp->ptr() )
-	{
-		return pp->ptr();
-	}
-	
-	return nullptr;
-}
-
-// Returns index of element with the specified name, or -1 if no such found
-int FindByName( HistCollect* pList, const unicode_t* name )
-{
-	for (int i = 0, count = pList->count(); i < count; i++)
-	{
-		if (unicode_is_equal( name, pList->get(i).data() ))
-		{
-			return i;
-		}
-	}
-	
-	return -1;
-}
-
-void AddFieldTextToHistory( const char* fieldName, const unicode_t* txt )
-{
-	if ( !HistoryCanSave() || !fieldName || !fieldName[0] || !txt || !txt[0] )
-	{
-		return;
-	}
-
-	std::vector<unicode_t> str = new_unicode_str( txt );
-
-	// get history collection for field
-	HistCollect* pList = HistGetList( fieldName );
-	if ( !pList)
-	{
-		clPtr<HistCollect> pUList = new HistCollect;
-		pList = pUList.ptr();
-		g_histHash[fieldName] = pUList;
-	}
-	
-	// check if item already exists in the list
-	const int index = FindByName( pList, str.data() );
-	if ( index != -1 )
-	{
-		// remove existing item
-		pList->del( index );
-	}
-	
-	// add item to the and of the list
-	pList->insert( 0, str );
-	
-	// limit number of elements in the list
-	while ( pList->count() > MAX_FIELD_HISTORY_COUNT )
-	{
-		pList->del( pList->count() - 1 );
-	}
 }
 
 static int uiClassNCEditLine = GetUiID( "NCEditLine" );
@@ -109,12 +30,12 @@ NCEditLine::NCEditLine( const char* fieldName, int nId, Win* parent, const unico
 				  rect )
 	, m_fieldName( fieldName )
 {
-	HistCollect* pList = HistGetList( m_fieldName );
+	HistCollect* pList = GetFieldHistCollect( m_fieldName );
 	if ( pList )
 	{
-		for ( int i = 0, count = pList->count(); i < count; i++ )
+		for ( int i = 0, count = pList->size(); i < count; i++ )
 		{
-			const unicode_t* u = pList->get( i ).data();
+			const unicode_t* u = pList->at( i ).data();
 			if ( u )
 			{
 				Append( u );
